# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DealerPro",
    "version": "v1"
  },
  "paths": {
...
    "/api/RegisterUser": {
      "post": {
        "tags": [
          "RegisterUser"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegistrationData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RegisterUser/register": {
      "post": {
        "tags": [
          "RegisterUser"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GoogleRegistrationData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "GoogleRegistrationData": {
      "type": "object",
      "properties": {
        "googleIdToken": {
          "type": "string"
        }
      }
    },
...
    "RegistrationData": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
  }
}